Exchange handler and controller in courses
